refactor: move valid func to new structure
